pyproject.toml
[project]
name = "FileAssistantAgent"
version = "0.1.0"
description = ""
authors = [
    {name = "kasrasadatsharifi",email = "kasrasadatsharifi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev-openai.dependencies]
openai-agents = "^0.0.16"

